requirements.txt
# R2 Data Challenge - Python Dependencies

# Core
pandas>=2.0.0
snowflake-connector-python[pandas]>=3.0.0

# Configuration
python-dotenv>=1.0.0

# dbt
dbt-core>=1.5.0
dbt-snowflake>=1.5.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Data generation
faker>=18.0.0
